Relative State Update
Fixed bad convergence by adding relative state function when training actor network. Relative state removes the absolute position from the state of the particle, allowing the network to evaluate/train regardless of in translational shifts of particle positions. Another improvement is to randomly select the relative center from the list of particles. This reduces the state space to train over.

train and test model functions are ready to use with new relative states